[refactor] board 리팩토링
