+++ src/pages/boards/[board_type]/index.tsx
@@ -1,59 +1,50 @@
-import PageTitle from '@src/components/Layout/PageTitle/PageTitle';
-import { useTheme } from 'styled-components';
-import SideBar from '@src/components/Layout/SideBar/SideBar';
 import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
-import PageWrapper from '@src/components/Layout/CommonCenterWrapper/CommonCenterWrapper';
+import styled from 'styled-components';
+import { useForm } from 'react-hook-form';
+import { AiOutlineSearch } from 'react-icons/ai';
 import { MdKeyboardArrowLeft, MdKeyboardArrowRight } from 'react-icons/md';
-import axios from 'axios';
-import {
-  Board,
-  Main,
-  PageButton,
-  PageButtonSection,
-  SearchIcon,
-  SearchInput,
-  Select,
-  SerachBar,
-  Table,
-  TableFooter,
-  TableHeader,
-  Tbody,
-  TdCenter,
-  Thead,
-  Tr,
-  WriteButton,
-} from './boards.style';
+import PageTitle from '@src/components/Layout/PageTitle/PageTitle';
+import SideBar from '@src/components/Layout/SideBar/SideBar';
+import PageWrapper from '@src/components/Layout/CommonCenterWrapper/CommonCenterWrapper';
+import DefaultButton from '@src/components/Button/Button';
+import Input from '@src/components/Input/Input';
+import Custom404 from '@src/pages/Error/404';
+import { PATH } from '@src/utils/urls';
 
 /**
  *  게시판 메인 페이지
  */
 
 function BoardPage() {
+  const { register } = useForm();
   const router = useRouter();
-  const { board_type } = router.query;
-  const onClickWrite = () => {
-    router.push(`/boards/${board_type}/new`);
-  };
-  const theme = useTheme();
+  const {
+    asPath,
+    query: { board_type },
+  } = router;
+
+  const boardType = Array.isArray(board_type) ? board_type[0] : board_type;
+  const validBoradPathList = Object.values(PATH.boards).map(
+    (board) => board.url,
+  );
+  const isValidPath = !boardType || !validBoradPathList.includes(asPath);
   const sideBarParmas = {
-    공지사항: '/boards/notice_board',
-    '정회원 게시판': '/boards/full_member_board',
-    '준회원 게시판': '/boards/associate_member_board',
+    [PATH.boards.notice.name]: PATH.boards.notice.url,
+    [PATH.boards.full.name]: PATH.boards.full.url,
+    [PATH.boards.graduate.name]: PATH.boards.graduate.url,
+    [PATH.boards.associate.name]: PATH.boards.associate.url,
   };
 
-  const [BoardData, SetBoardData] = useState([]);
-  // useEffect(() => {
-  //   axios
-  //     .get('/api/article/boards/notice_board/0/1')
-  //     .then((res) => SetBoardData(res.data));
-  // }, []);
+  const onClickWrite = () => {
+    router.push(`${asPath}/new`);
+  };
 
+  if (isValidPath) return <Custom404 />;
   return (
     <>
-      <PageTitle pageTitle="Boards"></PageTitle>
+      <PageTitle pageTitle={boardType}></PageTitle>
       <PageWrapper>
-        <SideBar menu_path={sideBarParmas}></SideBar>
+        <SideBar menu_path={sideBarParmas} />
         <Main>
           <Board>
             <TableHeader>
@@ -64,7 +55,10 @@ function BoardPage() {
                 <option value="3">ex2</option>
               </Select>
               <SerachBar>
-                <SearchInput placeholder="검색어를 입력해 주세요."></SearchInput>
+                <SearchInput
+                  register={register('search')}
+                  placeholder="검색어를 입력해 주세요."
+                ></SearchInput>
                 <SearchIcon size={20} />
               </SerachBar>
             </TableHeader>
@@ -130,5 +124,186 @@ function BoardPage() {
     </>
   );
 }
+
+export const Main = styled.div`
+  display: flex;
+`;
+
+export const Table = styled.table`
+  font-size: 1.6rem;
+  width: 100%;
+  margin-bottom: 1em;
+`;
+export const Board = styled.div`
+  width: 100%;
+`;
+export const Thead = styled.thead`
+  background-color: ${({ theme }) => theme.surfacePointDefault};
+  border-bottom: 1px solid ${({ theme }) => theme.borderDefault};
+  border-top: 1px solid ${({ theme }) => theme.borderDefault};
+  height: 2.4em;
+  line-height: 2.4em;
+`;
+export const Tbody = styled.tbody`
+  font-size: 1.6rem;
+  Tr {
+    cursor: pointer;
+    :hover {
+      transition: 0.2s;
+      background-color: ${({ theme }) => theme.surfacePointAlt};
+    }
+  }
+`;
+export const Tr = styled.tr`
+  height: 2.4em;
+  line-height: 2.4em;
+  border-bottom: 1px solid ${({ theme }) => theme.borderDefault};
+`;
+export const TdCenter = styled.td`
+  text-align: center;
+`;
+export const TableFooter = styled.div`
+  position: relative;
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+`;
+export const TableHeader = styled.div`
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 2.2em;
+  height: 35px;
+`;
+
+export const SearchInput = styled(Input)`
+  width: 100%;
+  height: 100%;
+  padding-left: 45px;
+`;
+export const SerachBar = styled.div`
+  display: flex;
+  width: 100%;
+  height: 100%;
+  position: relative;
+  align-items: center;
+`;
+export const SearchIcon = styled(AiOutlineSearch)`
+  position: absolute;
+  left: 15px;
+`;
+export const PageButton = styled.button`
+  width: 30px;
+  height: 30px;
+  border-radius: 50%;
+  background-color: inherit;
+  border: 1px solid ${({ theme }) => theme.borderDefault};
+  color: ${({ theme }) => theme.textWeek};
+  font-size: 1.6rem;
+  cursor: pointer;
+`;
+export const PageButtonSection = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 2em;
+`;
+export const WriteButton = styled(DefaultButton)`
+  align-self: flex-end;
+`;
+
+export const Form = styled.form`
+  position: relative;
+  height: calc(150vh - 70px - 50px); // header, footer 뺀 값
+  padding-bottom: 10vh;
+  box-sizing: border-box;
+  margin-top: 70px;
+  margin: 0 auto;
+  width: 700px;
+`;
+
+export const H1 = styled.h1`
+  margin-top: 0.5em;
+  margin-bottom: 0.5em;
+`;
+
+export const TitleInput = styled(Input)`
+  border: 0;
+  border-bottom: 0px solid ${({ theme }) => theme.borderDefault};
+  width: 100%;
+  padding-left: 15px;
+  font-size: 3rem;
+  height: 40px;
+  ::placeholder {
+    font-style: italic;
+    color: ${({ theme }) => theme.textWeek};
+  }
+`;
+
+export const CheckInput = styled.input`
+  align-self: flex-start;
+  margin-right: 1em;
+`;
+
+export const Options = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  font-size: 1.6rem;
+`;
+
+export const Header = styled.div`
+  display: flex;
+  align-items: center;
+  height: 40px;
+`;
+export const Select = styled.select`
+  background-color: inherit;
+  height: 100%;
+  color: ${({ theme }) => theme.textDefault};
+  font-size: 1.8rem;
+  margin-right: 1em;
+  width: 120px;
+  text-align: center;
+  border: 1px solid ${({ theme }) => theme.borderDefault};
+`;
+
+export const FileInput = styled.input`
+  display: none;
+`;
+export const FileInputLabel = styled.label`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 2rem;
+  border: 1px solid ${({ theme }) => theme.borderDefault};
+
+  border-radius: 4px;
+  height: 100px;
+  cursor: pointer;
+`;
+
+const TitleSection = styled.div`
+  margin-top: 2em;
+`;
+const EditorSection = styled.div`
+  margin-top: 2em;
+`;
+const OptionSection = styled.div`
+  margin-top: 2em;
+  padding: 24px;
+`;
+const FileSection = styled.div`
+  margin-top: 2em;
+  padding: 24px;
+`;
+const ButtonSection = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  margin-top: 2em;
+  padding: 24px;
+`;
+const Button = styled(DefaultButton)`
+  width: 100%;
+`;
 
 export default BoardPage;
